fix: Stop blocking when no longer modal due to resize
When the drawer is opened on small screens, it is in overlay mode where it is modal and blocks interaction with the rest of the app.

When the window is expanded enough, the drawer is no longer modal, but interaction with the app is still blocked. This is now fixed.